+++ src/midibus_portmidi.h
@@ -157,7 +157,7 @@ private:
     int m_queue;
 
     int m_ppqn;
-    int m_bpm;
+    double m_bpm;
 
     int  m_num_poll_descriptors;
     struct pollfd *m_poll_descriptors;
@@ -186,10 +186,10 @@ public:
     int get_num_out_buses();
     int get_num_in_buses();
 
-    void set_bpm(int a_bpm);
+    void set_bpm(double a_bpm);
     void set_ppqn(int a_ppqn);
 
-    int get_bpm()
+    double get_bpm()
     {
         return m_bpm;
     }
